Move getObjectInfo() to a separate header.
